[PLOP-64] Cut all ties with the CMS #1
Disable PHPStan for the "tests/"-folder. The only file there (a somewhat crude integration test) contains a problem that currently cannot be resolved to be in line with PHPStan. All other issues identified by PHPStan were resolved.

Allow StudyPortals\Template\Template to be called static in the test cases (this is needed as part of the original implementation).