[PLOP-206] Remove "local"-flag from Replace- and Condition-statements (#3)
This addresses an issue where using "local" as the comparison value in a condition-statement causes the parser to fail (as it assumes the "local"-flag is intended instead of comparing against the value "local"). There's no elegant way to solve this as that would break the set-comparison (i.e. "in"; "!in") functionality. Sets are not delimited, so the last element in the set could never be the world "local"...

Easiest way to solve this is to simply get rid off the "local"-replace/condition functionality altogether. It's used _once_ throughout the entirety of Portal/CMS.

The integration-tests were updated/verified as follows: Remove all "local"-flags from the integration-test in the master-branch and subsequently update the expected outcomes
of the test. If the change in this commit is proper, the expected outcomes should remain the same compare to the unchanged code of the master-branch.

Removing the "local"-properties from the Condition- and Replace-classes caused the serialised template format (Cache.bin) to change. Visual inspection of Cache.bin shows two sets of changes:

1. All "local"-properties are gone (they were set to boolean "false" in the reference file from master);
2. All "r"-elements (used to account for object references) are updated to account for the reduced number of serialised values (a direct result of removing the "local"-properties).