[PLOP-672] Implement Template::createStrict() (#5)
To prevent adding yet another static/global toggle to Template, a
slightly creative solution was required: The "strict" setting is stored
at the NodeTree-level (which is the lowest common denominator for what
Node::getRoot() returns).
In the process, Node::getRoot() was made more strict, it should always
return an instance of NodeTree (if it doesn't the template structure is
heavily b0rked).
Also, Node::display() used to swallow all exceptions raised while
generating the output. I *think* this is a remnant from the PHP5 times
when __toString() wasn't allowed to throw an exception. We might want to
test this assumption...

As Template inherits from NodeTree, it can modify the setting upon
creation, so no accessors are required for that.

From a performance perspective it is important to first do the
null-check and then the Node::getRoot()->isStrict() check (which
recursively walks the entire NodeTree).

For now there's only a generic StrictException, should more strict
checks be added later we can always extend from it.

The new StrictException raised by Condition::display() and
Replace::display() works wonderfully in PHP 7.4, but is not supported
in earlier PHP versions. So, a solution that works across the board (or
at least in all versions of PHP 7) is required.

Both the code and the accompanying integration-tests have been updated.
I've validated the code-paths in both PHP 7.2 and 7.4 (the tests cover
both).

For PHP < 7.4 we (mostly) reverted back to the old behaviour: Whenever
something goes wrong, the exception gets "eaten". An additional change
was made in the fact that we now always provide some feedback on the
error (instead of an empty string). This should aid in debugging
somewhat. The information provided is minimal, to ensure we don't
unintentionally leak anything sensitive.

The complex workaround with triggering an assertion in order to
determine whether the system was running in debug-mode was abandoned for
being opaque and overly complex.

For PHP >= 7.4 everything works wonderfully now. For PHP < 7.4 things
are okay, but still not optimal (as there's no way to actually signal
something went wrong in the __toString() method), but probably better
than providing a completely white screen...